remove custom tx
