+++ near/types/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod connector_args;
-pub mod transaction;
+pub mod bitcoin_connector_types;
 pub mod bitcoin_connector_events;
 pub mod mpc_types;
